refactor(split-rendering): extract layout, gutter, view_data (Phase 2 done)
Complete Phase 2 by extracting three more self-contained modules:

- layout.rs: split_layout, split_buffers_for_tabs, sync_viewport_to_content,
  resolve_view_preferences, calculate_view_anchor, calculate_compose_layout,
  calculate_viewport_end, render_separator + local types (SplitLayout,
  ViewPreferences, ViewAnchor, ComposeLayout)
- gutter.rs: render_left_margin with private LeftMarginContext,
  render_compose_margins
- view_data.rs: build_view_data + ViewData (orchestrates base_tokens,
  transforms, folding into the final Vec<ViewLine>)

None of these depend on any shared render-time mega struct.
All 37 split_rendering tests pass. mod.rs is down to 5,403 lines.
